EWPP-95: Use purl trait.
